Remove null-widget metadata to prevent bogus UI hiding
OSL compiler since Redshift 3.0.63 correctly handles null-widget metadata which causes parameters to be hidden in the UI (available only to script).  Some shaders in this repository were blindly using this, which until now was not a problem.

While this feature is legitimately useful sometimes, it does not seem appropriate for any of the current shaders.  Consequently, all occurrences have been removed.